Add unit tests for profile server
This commit adds unit tests for the profile server's upload and download HTTP handlers. It
uses mockery to handle the mocking of blob repository and event writing interfaces.

The worker's Work method has also been removed in favour of a single `event.Handler`
implementation so that the worker becomes easier to test.

Signed-off-by: David Bond <davidsbond93@gmail.com>