Rework client profile method
This commit modifies the client's Profile method to now perform both profiling and
uploading in a single step. This appeared to be better to me when adding tests for
the client's upload, download & profile methods. It didn't make a lot of sense to
me for the client to have a method to grab a profile and return it rather than just
perform the upload afterwards since the client already has all the information
required to do that.

This also makes the scraper code a little simpler.

Signed-off-by: David Bond <davidsbond93@gmail.com>