Rework download command parameters
This commit modifies how the `download` command works to no longer pipe to
stdout but to instead use a `--output` or `-o` flag with a default value of
`default.pgo`.

This is done because if you use `>>` with logging enabled your profile ends
up corrupted with any logs generated interlaced with the profile.

Signed-off-by: David Bond <davidsbond93@gmail.com>