Add command to delete profile
This commit adds a new `delete` command to the CLI which will delete any merged or
pending profiles for an application.

It also reworks the `Blob.List` method to use an `iter.Seq2` iterator. This seemed
like a good place to use it as the underlying gocloud.dev package uses a pre
native iterator already, so this saves an additional loop.

Signed-off-by: David Bond <davidsbond93@gmail.com>