Rework scraper configuration (#7)
This commit reworks how the scraper is configured to make it focused on having a single
application per scraper. The app value is now set at the top level and all targets
profiled by the scraper will be uploaded under a single app.

The reason for this change is that before this, sampling wasn't done per
application, so if the scraper was set up with multiple applications there was
no way to control sampling per-app. Instead, all targets share the same
sample size, which could end up with some apps being sampled more than others.

Rather than add a complicated way of sampling per app, I've just made the scraper
responsible for a single app.

Signed-off-by: David Bond <davidsbond93@gmail.com>