Move clean command to the client side (#9)
This commit undoes everything I did already with clean. I'm not sure why I went down
the rabbit hole of adding a new HTTP endpoint and a bunch of extra functions and
helpers when the CLI could just use list and delete itself. This seems like a
much simpler way to do it.

Signed-off-by: David Bond <davidsbond93@gmail.com>