Initial consul support
This commit adds initial support for HashiCorp consul as a source of scrapable
targets for the scraper. It is confiugred using the typical environment variables
one would use for a consul client. This commit also includes an integration test
against a running consul instance.

It functions pretty much the same as the nomad source in terms of logic.

Signed-off-by: David Bond <davidsbond93@gmail.com>